Inline worker service client surface
