+++ src/refiner/platform/client/api.py
@@ -226,6 +226,45 @@ class MacrodataClient:
             json_payload=request.to_dict(),
         )
 
+    def start_worker_services(
+        self,
+        *,
+        job_id: str,
+        stage_index: int,
+        worker_id: str,
+        services: list[dict[str, Any]],
+    ) -> dict[str, Any]:
+        response_data = request_json(
+            method="POST",
+            path=f"/api/jobs/{job_id}/stages/{stage_index}/workers/{worker_id}/services/start",
+            api_key=self.api_key,
+            base_url=self.base_url,
+            json_payload={"services": services},
+            timeout_s=10.0,
+        )
+        if not isinstance(response_data, dict):
+            raise ValueError("runtime services response must be a JSON object")
+        return response_data
+
+    def get_worker_service(
+        self,
+        *,
+        job_id: str,
+        stage_index: int,
+        worker_id: str,
+        service_id: str,
+    ) -> dict[str, Any]:
+        response_data = request_json(
+            method="GET",
+            path=f"/api/jobs/{job_id}/stages/{stage_index}/workers/{worker_id}/services/{service_id}",
+            api_key=self.api_key,
+            base_url=self.base_url,
+            timeout_s=10.0,
+        )
+        if not isinstance(response_data, dict):
+            raise ValueError("runtime service response must be a JSON object")
+        return response_data
+
 
 __all__ = [
     "MacrodataApiError",
